moved the generateTransportToken from the controller to the cert util since it was a strange pattern to import the function from a controller
