+++ src/controllers/auth.ts
@@ -3,6 +3,7 @@ import { success, failure } from '../utils/res'
 import { loadConfig } from '../utils/config'
 import * as rsa from '../crypto/rsa'
 import * as tribes from '../utils/tribes'
+import { generateTransportTokenKeys } from '../utils/cert'
 import * as fs from 'fs'
 import { Req } from '../types'
 
@@ -86,14 +87,4 @@ export async function requestTransportKey(req: Req, res) {
 
   const transportTokenKeys = await generateTransportTokenKeys()
   success(res, { transport_key: transportTokenKeys })
-}
-
-export async function generateTransportTokenKeys() {
-  const transportTokenKeys: { [k: string]: string } = await rsa.genKeys()
-  fs.writeFileSync(config.transportPublicKeyLocation, transportTokenKeys.public)
-  fs.writeFileSync(
-    config.transportPrivateKeyLocation,
-    transportTokenKeys.private
-  )
-  return transportTokenKeys.public
 }
